* update code standards
