+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/rules/RulesEngineDateUtil.java
@@ -6,7 +6,6 @@ import android.text.TextUtils;
 import com.vijay.jsonwizard.utils.FormUtils;
 import com.vijay.jsonwizard.utils.Utils;
 
-import org.joda.time.DateTime;
 import org.joda.time.Days;
 import org.joda.time.LocalDate;
 import org.joda.time.Months;
